backend/requirements.txt
flask==3.1.2
flask-socketio==5.5.1
eventlet==0.40.3
flask-cors==6.0.1
pydantic==2.11.7
# CTP API dependencies
ctp-python==6.7.7
numpy==1.24.3
pandas==2.0.3
# Additional utilities
python-dotenv==1.0.0
loguru==0.7.2
